[18USA] Create a common module for resource track handling
